refactor(notebook): share runtime surface components
