+++ apps/elements/app/global.css
@@ -16,7 +16,6 @@
 @source "../../../src/components/rail/**/*.{ts,tsx}";
 @source "../../../src/components/ui/**/*.{ts,tsx}";
 @source "../../../src/components/widgets/**/*.{ts,tsx}";
-@source "../../../apps/notebook/src/components/**/*.{ts,tsx}";
 @source "../../../packages/sift/src/**/*.{ts,tsx}";
 
 :root {
