+++ apps/notebook/src/components/runtime-surface-types.ts
@@ -1,8 +1,7 @@
-export type { PyProjectDeps, PyProjectInfo } from "runtimed";
-export type { TrustInfo, TyposquatWarning } from "@nteract/notebook-host";
-
-export type EnvSyncState =
-  | { status: "not_running" }
-  | { status: "not_uv_managed" }
-  | { status: "synced" }
-  | { status: "dirty"; added: string[]; removed: string[] };
+export type {
+  EnvSyncState,
+  PyProjectDeps,
+  PyProjectInfo,
+  TrustInfo,
+  TyposquatWarning,
+} from "@/components/notebook/runtime-surface-types";
