+++ src/components/notebook/RuntimeDecisionDialog.tsx
@@ -1,0 +1,49 @@
+import type { ReactNode } from "react";
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from "@/components/ui/dialog";
+
+export interface RuntimeDecisionDialogProps {
+  open: boolean;
+  onOpenChange: (open: boolean) => void;
+  title: ReactNode;
+  description: ReactNode;
+  icon: ReactNode;
+  children: ReactNode;
+  footer: ReactNode;
+  testId?: string;
+}
+
+export function RuntimeDecisionDialog({
+  open,
+  onOpenChange,
+  title,
+  description,
+  icon,
+  children,
+  footer,
+  testId = "runtime-decision-dialog",
+}: RuntimeDecisionDialogProps) {
+  return (
+    <Dialog open={open} onOpenChange={onOpenChange}>
+      <DialogContent className="max-w-lg" showCloseButton={false} data-testid={testId}>
+        <DialogHeader>
+          <DialogTitle className="flex items-center gap-2">
+            {icon}
+            {title}
+          </DialogTitle>
+          <DialogDescription>{description}</DialogDescription>
+        </DialogHeader>
+
+        {children}
+
+        <DialogFooter className="gap-2 sm:gap-0">{footer}</DialogFooter>
+      </DialogContent>
+    </Dialog>
+  );
+}
